requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.4.0
psycopg2-binary>=2.9.0